Core: ----- * add support for disabling configuration inheritance to child processes * module configuration settings are stored in MODREF structures instead of IMTE - single module can have different configurations under different processes (better accuracy, overriding is supported, recursive inheritance is supported, speed gains) * updated SDK with missing resolver flags
Compatibility tab:
------------------
* added option to disable configuration inheritance to child processes
* added option to override module configuration with process'es configuration
* changed options descriptions
* added baloon tips
* fixed COM implementation

Setup:
------
* use kexcrt instead of msvcrt
* forbid direct downgrades
* updates for IMTE->MODREF changes

+++ sheet/server.h
@@ -22,7 +22,7 @@
 #ifndef _SERVER_H
 #define _SERVER_H
 
-extern UINT g_LockCount;
+extern long g_LockCount;
 extern HMODULE g_hModule;
 
 #endif // _SERVER_H
